backend/chatbot/requirements.txt
fastapi
uvicorn
transformers
torch
accelerate
sentence-transformers
chromadb
pydantic
pydantic-settings
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.27.0
slowapi>=0.1.9
google-generativeai
google-generativeai
